Qt Creator v4.1.0

Conflicts:
	dist/gdb/Makefile.linux
	dist/gdb/Makefile.mingw
	dist/gdb/Makefile.osx
	doc/src/qtcreator.qdoc
	src/plugins/projectexplorer/gcctoolchain.cpp
	src/plugins/qmakeprojectmanager/makestep.cpp
	src/plugins/qmakeprojectmanager/qmakeprojectmanagerplugin.cpp
	src/plugins/qtsupport/exampleslistmodel.cpp
	src/plugins/qtsupport/gettingstartedwelcomepage.cpp

+++ src/libs/utils/smallstringliteral.h
@@ -90,12 +90,12 @@ public:
 
     const_reverse_iterator rbegin() const noexcept
     {
-        return const_reverse_iterator(end() - 1l);
+        return const_reverse_iterator(end() - static_cast<std::size_t>(1));
     }
 
     const_reverse_iterator rend() const noexcept
     {
-        return const_reverse_iterator(begin() - 1l);
+        return const_reverse_iterator(begin() - static_cast<std::size_t>(1));
     }
 
     constexpr static
@@ -134,4 +134,3 @@ private:
 
 #pragma pop_macro("noexcept")
 #pragma pop_macro("constexpr")
-
